refactor: rename spell resistanceCheck to resistedBy and settle the enum
The shared spell field added in #757/#758 covered only POW-vs-POW resistance
rolls, but it now also has to describe spirit-combat gates, which aren't
"resistance checks". Rename `resistanceCheck` -> `resistedBy` and
`SpellResistanceCheckEnum` -> `SpellResistedByEnum`.

Declare the full enum now so pack content is authored once:
none | resistanceRoll | resistanceRollArea | resistanceRollPerTarget |
spiritCombatRound | spiritCombatDefeat. Only none and resistanceRoll (the old
"single") are implemented; maybePromptResistanceRollForCast keeps its guard and
treats the other four as "no automated step" until their own issues wire them
up. Two spirit-combat values because RQG has two distinct gates - one won round
(willing target, e.g. Bind Ghost) vs. driven to 0 MP (e.g. Spirit Binding).

No behaviour change and no migration (nothing released). Turn Undead's
POW-vs-magic-points stays out - it's a future orthogonal resistedWith field.

Closes #1068
