+++ buildScripts/i18n-dynamic-key-map.ts
@@ -34,7 +34,7 @@ import {
   SpellRangeEnum,
   SpellDurationEnum,
   SpellConcentrationEnum,
-  SpellResistanceCheckEnum,
+  SpellResistedByEnum,
 } from "../src/data-model/item-data/spell";
 import { AbilitySuccessLevelEnum } from "../src/rolls/ability-roll/ability-roll.defs";
 
@@ -147,7 +147,7 @@ export const dynamicKeyMap: Record<string, readonly string[]> = {
     "undefined",
   ],
   "RQG.Item.Spell.RangeEnum.": [...Object.values(SpellRangeEnum).filter(Boolean), "undefined"],
-  "RQG.Item.Spell.ResistanceCheckEnum.": [...Object.values(SpellResistanceCheckEnum)],
+  "RQG.Item.Spell.ResistedByEnum.": [...Object.values(SpellResistedByEnum)],
 
   // Item - Weapon
   "RQG.Item.Weapon.combatManeuver.": [...combatManeuverNames],
