+++ src/data-model/item-data/rune-magic-data-model.ts
@@ -2,7 +2,7 @@ import type { RqgActor } from "@actors/rqg-actor.ts";
 import type { RqgItem } from "@items/rqg-item.ts";
 import { RqgItemDataModel } from "./rqg-item-data-model";
 import { migrateSpellBooleanFields, spellSchemaFields } from "../shared/spell-schema-fields";
-import { maybePromptResistanceRollForCast } from "../shared/spell-resistance-check";
+import { maybePromptResistanceRollForCast } from "../shared/spell-resisted-by";
 import type { RqidLink } from "../shared/rqid-link";
 import type { RqidString } from "../../system/api/rqid-api";
 import { rqidLinkArraySchemaField } from "../shared/rqid-link-field";
@@ -353,7 +353,7 @@ export class RuneMagicDataModel extends RqgItemDataModel<RuneMagicSchema, { chan
     );
 
     await maybePromptResistanceRollForCast(
-      this.resistanceCheck,
+      this.resistedBy,
       runeMagicRoll.successLevel,
       casterActor,
       token,
