+++ src/items/rune-magic-item/rune-magic-sheet-v2.ts
@@ -1,7 +1,7 @@
 import { ItemTypeEnum } from "@item-model/item-types.ts";
 import { getSelectRuneOptions, isDocumentSubType } from "../../system/util";
 import { RqgItemSheetV2, type RqgItemSheetContext } from "../rqg-item-sheet-v2";
-import { SpellDurationEnum, SpellRangeEnum, SpellResistanceCheckEnum } from "@item-model/spell.ts";
+import { SpellDurationEnum, SpellRangeEnum, SpellResistedByEnum } from "@item-model/spell.ts";
 import { systemId } from "../../system/config";
 import { templatePaths } from "../../system/load-handlebars-templates";
 import type { CultItem } from "@item-model/cult-data-model.ts";
@@ -12,7 +12,7 @@ interface RuneMagicSheetContext extends RqgItemSheetContext {
   allRuneOptions: SelectOptionData<string>[];
   rangeOptions: SelectOptionData<SpellRangeEnum>[];
   durationOptions: SelectOptionData<SpellDurationEnum>[];
-  resistanceCheckOptions: SelectOptionData<SpellResistanceCheckEnum>[];
+  resistedByOptions: SelectOptionData<SpellResistedByEnum>[];
   actorCultOptions: SelectOptionData<string>[];
 }
 
@@ -59,9 +59,9 @@ export class RuneMagicSheetV2 extends RqgItemSheetV2 {
         value: range,
         label: "RQG.Item.Spell.DurationEnum." + (range || "undefined"),
       })),
-      resistanceCheckOptions: Object.values(SpellResistanceCheckEnum).map((value) => ({
+      resistedByOptions: Object.values(SpellResistedByEnum).map((value) => ({
         value: value,
-        label: "RQG.Item.Spell.ResistanceCheckEnum." + value,
+        label: "RQG.Item.Spell.ResistedByEnum." + value,
       })),
       actorCultOptions: this.getActorCultOptions(),
       allRuneOptions: getSelectRuneOptions("RQG.Item.RuneMagic.AddRuneMagicRunePlaceholder"),
